Moved utils to their own file.
